Move the CNAME
Every time we run `middleman deploy`, it pushes over whatever's on the
master branch. That means the CNAME was being removed with every deploy.
The previous commit
https://github.com/refreshthetriangle/refreshthetriangle.github.io/commit/eb6333c57aee2b447297ea96e62dedf530a881b9
tried putting the CNAME at the root of the source branch, but that
wasn't included with the build from `middleman deploy`.

Instead, we should be able to move it to the `/source` directory, which
will be included at master's root in all future deployments.